% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbdemux2}
\alias{bbdemux2}
\title{Demultiplex fusion primers using BBmap Seal}
\usage{
bbdemux2(
  install = NULL,
  fwd,
  rev = NULL,
  Fbarcodes = NULL,
  Rbarcodes = NULL,
  names = NULL,
  restrictleft = NULL,
  out.dir = "demux",
  kmer = NULL,
  hdist = 0,
  degenerate = TRUE,
  force = TRUE,
  mem = NULL,
  threads = NULL,
  quiet = FALSE
)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{fwd}{(Required) Vector of locations of forward reads}

\item{rev}{(Optional) Vector of locations of reverse reads}

\item{Fbarcodes}{(Required) Barcodes used in forward reads}

\item{Rbarcodes}{(Optional) Barcodes used in reverse reads}

\item{restrictleft}{(Optional) Defaults to the size of the largest primer.
Restricts the kmer search for primer sequences to just the left side of the molecule.}

\item{out.dir}{(Optional) Default "demux"
The path to write the output reads.}

\item{kmer}{(Optional) default the size of the smallest primer will be used.
The kmer size to use for primer searching.}

\item{hdist}{(Optional) Default = 0. The hamming distance (number of substitution errors) allowed for mismatch to the query primer.}

\item{degenerate}{(Optional) Default TRUE.
Option to search for all possible primer combinations for degenerate primers}

\item{force}{(Optional) Default TRUE
Option to overwrite existing output files.}

\item{mem}{(Optional) Default autodetect
GB of memory to use
WARNING: mem detection can fail on cluster computing currently}

\item{threads}{(Optional) Default autodetect
Number of CPU threads to use
WARNING: Thread detection can fail on cluster computing currently}

\item{interleaved}{(Optional) Default FALSE
Option to input interleaved reads}
}
\value{

}
\description{
Demultiplex fusion primers using BBmap Seal
}
