% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbsplit2}
\alias{bbsplit2}
\title{Split interleaved reads}
\usage{
bbsplit2(install = NULL, file, force = FALSE, quiet = FALSE)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{force}{(Optional) Default TRUE Option to overwrite existing output files.}

\item{files}{(Required) Vector of locations of interleaved read files to split}
}
\description{
Split interleaved reads
}
