% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbmap_utilities.R
\name{bbtrim2}
\alias{bbtrim2}
\title{Trim primers using BBDuk}
\usage{
bbtrim2(
  install = NULL,
  fwd,
  rev = NULL,
  primers,
  checkpairs = FALSE,
  restrictleft = NULL,
  out.dir = NULL,
  trim.end = "left",
  ordered = TRUE,
  kmer = NULL,
  mink = FALSE,
  tbo = TRUE,
  tpe = TRUE,
  hdist = 0,
  degenerate = TRUE,
  force = TRUE,
  maxlength = NULL,
  quiet = FALSE
)
}
\arguments{
\item{install}{(Required) Install location for bbmap}

\item{fwd}{(Required) Vector of locations of forward reads}

\item{rev}{(Optional) Vector of locations of reverse reads}

\item{primers}{(Required) Forward and reverse primers to trim}

\item{checkpairs}{(Optional) Whether paired end checking should be conducted}

\item{restrictleft}{(Optional) Defaults to the size of the largest primer.
Restricts the kmer search for primer sequences to just the left side of the molecule.}

\item{out.dir}{(Optional) Default "trimmed"
The path to write the output reads.}

\item{trim.end}{(Optional) Default is "left"
End of the molecule to trim primers from. "left" will trim primers from the 3' end of both forward and reverse reads.
Change to "right" only if the amplicon was too short and the sequencer has read into the other end of the molecule.}

\item{ordered}{(Optional) Default TRUE
Set to TRUE to output reads in same order as input.}

\item{kmer}{(Optional) default the size of the smallest primer will be used.
The kmer size to use for primer searching.}

\item{mink}{(optional) Default FALSE
Look for shorter kmers at read tips down to this length}

\item{tpe}{(Otional) Default TRUE
Trim pairs evenly. When kmer right-trimming, trim both reads to the minimum length of either.}

\item{hdist}{(Optional) Default 0. The hamming distance (number of substitution errors) allowed for mismatch to the query primer.}

\item{degenerate}{(Optional) Default TRUE.
Option to search for all possible primer combinations for degenerate primers}

\item{force}{(Optional) Default TRUE
Option to overwrite existing output files.}

\item{maxlength}{(Optional) Default FALSE
Remove all reads above a maximum length. Useful for removing reads where no primers were found.}
}
\description{
Trim primers using BBDuk
}
