% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{cluster_otus}
\alias{cluster_otus}
\title{OTU clustering of DADA2 seqtab}
\usage{
cluster_otus(
  seqtab,
  method = "complete",
  similarity = 0.97,
  rename = FALSE,
  cores = 1
)
}
\arguments{
\item{seqtab}{A DADA2 generated ASV table with columns as sequences and rows as samples}

\item{method}{An agglomeration method to parse to DECIPHER::IdClusters.
This should be (an abbreviation of) one of "complete", "single", "UPGMA", "WPGMA", "NJ", "ML", or "inexact".
(See help page on DECIPHER::IdClusters for more information.)}

\item{similarity}{A similarity to cluster at}

\item{rename}{Whether clusters should be renamed to OTU<cluster #>}

\item{cores}{The number of processsor cores to use}
}
\value{

}
\description{
Thanks to Michael mclaren and Erik Wright for code
}
