% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescing join}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::full_join,
  ...
)
}
\arguments{
\item{x}{The primary data frame or tibble from which values will be prefered}

\item{y}{The secondary data frame or tibble from which values will only be used when values in x are NA}

\item{by}{A character vector of columns to join by.}

\item{suffix}{The suffixes from the dplyr join to remove.}

\item{join}{The type of dplyr join to perform}

\item{...}{}
}
\description{
A join to combine datasets containing identical non-key columns in varying states of completeness
}
