% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_tracking.R
\name{create_logsheet}
\alias{create_logsheet}
\title{Create logsheet}
\usage{
create_logsheet(SampleSheet, runParameters)
}
\arguments{
\item{SampleSheet}{A samplesheet or set of samplesheets in MiSeq or NovaSeq format}

\item{runParameters}{A runParamers.xml or set of these output by the sequencer}
}
\description{
Create logsheet
}
