% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{estimate_truclen}
\alias{estimate_truclen}
\title{Automatically estimate trunclen for filtering}
\usage{
estimate_truclen(
  fwd,
  rev = NULL,
  threshold = 25,
  maxlength,
  minlength,
  qa_sample = 5e+05,
  overlap_sample = 100,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Automatically estimate trunclen for filtering
}
