% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lowest_classified}
\alias{lowest_classified}
\title{Get the lowest classified ranks for the OTUs}
\usage{
lowest_classified(x, sep = "__", return = "rank", rownames = TRUE)
}
\arguments{
\item{x}{A matrix, data frame or phyloseq object}

\item{sep}{The seperator between the rank, and the name
i.e. G__Drosophila would be the default for an OTU that could only be assigned to the genus Drosophila}

\item{return}{The type of values to return.
Options include 'rank' which returns a character vector listing the lowest taxonomic each otu was classified to, i.e. Species, Genus etc
'taxon' returns a character vector listing the lowest taxonomic name each otu was classified to i.e. Drosophila_suzukii, Diptera etc
'both' returns a data frame containing both of the above}

\item{rownames}{Whether rownames should be returned, or instead an OTU column}
}
\value{

}
\description{
Get the lowest classified ranks for the OTUs
}
\seealso{
na_to_unclassified unclassified_to_na
}
