% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{n_overlap}
\alias{n_overlap}
\title{Number of bases overlapping between forward and reverse reads}
\usage{
n_overlap(fwd, rev, sample = 100)
}
\arguments{
\item{sample}{}
}
\value{

}
\description{
Number of bases overlapping between forward and reverse reads
}
