% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{na_to_unclassified}
\alias{na_to_unclassified}
\title{Convert NA classifications to the their highest successful classification}
\usage{
na_to_unclassified(x, sep = "__", rownames = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{A matrix, data frame or phyloseq object}

\item{sep}{The seperator between the rank, and the name
i.e. G__Drosophila would be the default for an OTU that could only be assigned to the genus Drosophila}

\item{rownames}{Whether rownames should be returned, or instead an OTU column}

\item{quiet}{Whether progress should be printed to console}
}
\value{

}
\description{
Convert NA classifications to the their highest successful classification
}
\seealso{
unclassified_to_na
}
