% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{plot_lengths}
\alias{plot_lengths}
\title{Plot readlengths of fastq}
\usage{
plot_lengths(dir, aggregate = FALSE, sample = 1e+06)
}
\arguments{
\item{dir}{A directory of fastq files}

\item{aggregate}{Whether plot should be faceted by sample or aggregated together}

\item{sample}{How many reads to sample from each fastq, default 1e6}
}
\value{

}
\description{
Plot readlengths of fastq
}
