% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{proportions}
\alias{proportions}
\title{Convert phyloseq table to proportions
#'}
\usage{
proportions(x, thresh = NA, na_rm = FALSE, ...)
}
\arguments{
\item{x}{An OTU Table from Phyloseq}

\item{thresh}{A minimum abundance threshold to filter taxa}

\item{na_rm}{Whether NA's should be removed}

\item{...}{}
}
\value{

}
\description{
Convert phyloseq table to proportions
#'
}
