% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ps_to_dnabin}
\alias{ps_to_dnabin}
\title{ps_to_dnabin}
\usage{
ps_to_dnabin(ps, seqnames = "unique")
}
\arguments{
\item{ps}{A \code{phyloseq} object that contains \code{\link[phyloseq]{refseq}}.
If there the \code{refseq} slot is not filled, this function will try pull the
sequences from \code{\link[phyloseq]{get_taxa}}}

\item{seqnames}{(optional) A taxonomic rank from the \code{\link[phyloseq]{tax_table}} which will be used to name the sequences.
alternatively "unique" will uniquely name the sequences as \code{ASV_#},
"sequence" will name the sequences in the fasta exactly the same same as the sequence, usfull for LULU curation.}
}
\description{
ps_to_dnabin
}
\examples{
save_fasta(ps)
save_fasta(ps = ps, file = "sequences.fasta", rank = "Genus")
}
