% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{summarise_index}
\alias{summarise_index}
\title{Summarise index}
\usage{
summarise_index(fq, qualityType = "Auto")
}
\arguments{
\item{fq}{A path to a fastq file}

\item{qualityType}{The quality encoding of the fastq file}
}
\value{

}
\description{
Summarise index
}
