% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{align_overlap}
\alias{align_overlap}
\title{Minimum overlap between a query and reference}
\usage{
align_overlap(query, ref)
}
\arguments{
\item{ref}{}
}
\description{
Minimum overlap between a query and reference
}
