% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utilities.R
\name{n_overlap}
\alias{n_overlap}
\title{Number of bases overlapping between forward and reverse reads}
\usage{
n_overlap(fwd, rev, sample = 100)
}
\arguments{
\item{fwd}{A path to a fastq file containing forward reads}

\item{rev}{A path to a fastq file containing reverse reads}

\item{sample}{The number of reads to sample from the fastq files}
}
\description{
Number of bases overlapping between forward and reverse reads
}
